[Chore] app.ts 위치변경
전: back/app.ts
후: back/src/app.ts
